# docs/chatgpt-gpt/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MERKEN Wordbook API for ChatGPT",
    "description": "ChatGPTの会話からMERKENの単語帳に単語を追加し、苦手な単語(クイズで間違いが多い単語)の取得や、文法・語法問題集(Vintage型: 空欄補充・英語4択・解説つき)の作成・出題を行うためのAPI。訳・例文・選択肢・問題・解説はすべてChatGPT側で生成して送信する(サーバー側ではAI生成を行わない)。単語帳(wordbook)と語法問題集(grammar book)は別のフォーマットであり、混合してはならない。",
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "https://www.merken.jp"
    }
  ],
  "paths": {
    "/api/chatgpt/grammar-books": {
      "get": {
        "operationId": "listGrammarBooks",
        "summary": "文法・語法問題集の一覧を取得する",
        "description": "語法問題を追加する前に呼び出し、追加先の問題集(bookId)を決める。単語帳(listWordbooks)とは別物。",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "問題集一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "books": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "title": { "type": "string" },
                          "updatedAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createGrammarBook",
        "x-openai-isConsequential": false,
        "summary": "新しい文法・語法問題集を作成する",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "additionalProperties": false,
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120,
                    "description": "問題集のタイトル(例: 語法 動詞編)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "作成された問題集",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "book": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "title": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chatgpt/grammar-questions": {
      "get": {
        "operationId": "getGrammarQuestions",
        "summary": "問題集の語法問題を取得する",
        "description": "保存済みの語法問題を取得する。ChatGPT上で出題・答え合わせ・解説をするときに使う。",
        "parameters": [
          {
            "name": "bookId",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "問題リスト",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "questions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "sentence": { "type": "string" },
                          "choices": { "type": "array", "items": { "type": "string" } },
                          "correctIndex": { "type": "integer" },
                          "explanation": { "type": "string" },
                          "grammarPoint": { "type": "string" },
                          "sentenceJa": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addGrammarQuestions",
        "x-openai-isConsequential": false,
        "summary": "問題集に文法・語法問題を追加する",
        "description": "Vintage型の語法問題を追加する。問題文は空欄マーカー ___ を必ず含み、選択肢は英語4つ(文法・語法の観点で紛らわしいもの)、各問題に解説(explanation)が必須。単語学習用の単語はここに入れず addWords を使うこと。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["bookId", "questions"],
                "additionalProperties": false,
                "properties": {
                  "bookId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "追加先の問題集ID(listGrammarBooks / createGrammarBookで取得)"
                  },
                  "questions": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 50,
                    "items": {
                      "type": "object",
                      "required": ["sentence", "choices", "correctIndex", "explanation"],
                      "additionalProperties": false,
                      "properties": {
                        "sentence": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 300,
                          "description": "空欄マーカー ___ を含む英語の問題文(例: She insisted that he ___ the meeting.)"
                        },
                        "choices": {
                          "type": "array",
                          "minItems": 4,
                          "maxItems": 4,
                          "items": { "type": "string", "minLength": 1, "maxLength": 80 },
                          "description": "英語の選択肢4つ(重複不可)。文法・語法の観点で紛らわしい動詞の形・関係代名詞・前置詞などを含める。日本語は使わない"
                        },
                        "correctIndex": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 3,
                          "description": "正解のインデックス(0始まり)"
                        },
                        "explanation": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1000,
                          "description": "必須。なぜ正解なのか+主要な誤答がなぜ違うのかを日本語で簡潔に解説する(Vintageの解説のイメージ)"
                        },
                        "grammarPoint": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 40,
                          "description": "文法項目タグ(例: 仮定法現在, 関係代名詞, 分詞構文)"
                        },
                        "sentenceJa": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 300,
                          "description": "問題文の日本語訳(任意)"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "追加結果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "addedCount": { "type": "integer" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Pro限定機能(code: PRO_REQUIRED)または問題集へのアクセス権なし"
          }
        }
      }
    },
    "/api/chatgpt/struggling-words": {
      "get": {
        "operationId": "getStrugglingWords",
        "summary": "ユーザーの苦手な単語(間違いが多い単語)を取得する",
        "description": "ユーザーがMERKENのクイズで間違えた単語を、誤答回数が多い順に返す。「苦手な単語で練習したい」「よく間違える単語を復習させて」等の依頼で呼び出し、返ってきた単語を使ってクイズ・例文作成・会話練習を行う。",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "誤答回数の多い順の単語リスト",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "totalCount": {
                      "type": "integer",
                      "description": "集計対象となった苦手単語の総数"
                    },
                    "words": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "english": { "type": "string" },
                          "japanese": { "type": "string" },
                          "missCount": {
                            "type": "integer",
                            "description": "クイズでの誤答回数"
                          },
                          "lastMissedAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "最後に間違えた日時"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Pro限定機能(code: PRO_REQUIRED)"
          }
        }
      }
    },
    "/api/chatgpt/projects": {
      "get": {
        "operationId": "listWordbooks",
        "summary": "ユーザーの単語帳一覧を取得する",
        "description": "単語を追加する前に呼び出し、追加先の単語帳(projectId)を決める。",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "単語帳一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "projects": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "title": { "type": "string" },
                          "updatedAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createWordbook",
        "x-openai-isConsequential": false,
        "summary": "新しい単語帳を作成する",
        "description": "追加先に適した単語帳がない場合に作成する。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "additionalProperties": false,
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120,
                    "description": "単語帳のタイトル(例: ChatGPTで学んだ単語)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "作成された単語帳",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "project": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "title": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chatgpt/words": {
      "post": {
        "operationId": "addWords",
        "x-openai-isConsequential": false,
        "summary": "単語帳に単語を追加する",
        "description": "会話に出てきた英単語をユーザーの単語帳に追加する。日本語訳(japanese)は必須。例文・発音・品詞・クイズ用の誤答選択肢(distractors)もChatGPTが生成して必ず含めること。サーバー側では翻訳や例文の自動生成は行われない。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["words"],
                "additionalProperties": false,
                "properties": {
                  "words": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 100,
                    "items": {
                      "type": "object",
                      "required": ["projectId", "english", "japanese"],
                      "additionalProperties": false,
                      "properties": {
                        "projectId": {
                          "type": "string",
                          "format": "uuid",
                          "description": "追加先の単語帳ID(listWordbooks / createWordbookで取得)"
                        },
                        "english": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200,
                          "description": "英単語または熟語(辞書形)"
                        },
                        "japanese": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 300,
                          "description": "簡潔な日本語訳(必須。会話の文脈に合う訳を選ぶ)"
                        },
                        "exampleSentence": {
                          "type": "string",
                          "maxLength": 500,
                          "description": "その単語を使った自然な英語の例文"
                        },
                        "exampleSentenceJa": {
                          "type": "string",
                          "maxLength": 500,
                          "description": "例文の日本語訳"
                        },
                        "pronunciation": {
                          "type": "string",
                          "maxLength": 120,
                          "description": "IPA発音記号"
                        },
                        "partOfSpeechTags": {
                          "type": "array",
                          "maxItems": 10,
                          "items": { "type": "string", "minLength": 1, "maxLength": 32 },
                          "description": "品詞タグ(noun, verb, adjective, adverb, idiom など英語表記)"
                        },
                        "distractors": {
                          "type": "array",
                          "maxItems": 10,
                          "items": { "type": "string", "minLength": 1, "maxLength": 300 },
                          "description": "4択クイズ用の誤答選択肢3つ。正解とは別の英単語の日本語訳から作ること。出題語自身の別の意味は使わない"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "追加結果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "addedCount": { "type": "integer" },
                    "words": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "projectId": { "type": "string" },
                          "english": { "type": "string" },
                          "japanese": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Pro限定機能(code: PRO_REQUIRED)または単語帳へのアクセス権なし"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {}
  }
}